+++ src/pages/player/detail/CoachDetailPage.tsx
