+++ src/pages/player/detail/PlayerDetailPage.tsx
